refactor: move files to appropriate folders
